refactor: split infoforms store and template parsing into own files
